Permanent host status refresh every minute

+++ Code/commands/index.js
@@ -18,7 +18,7 @@ client.on('message', message => {
         if (Manager.usersGame('<@' + message.author.id + '>') != null && Manager.usersGame('<@' + message.author.id + '>').channel[Manager.usersGame('<@' + message.author.id + '>').players.indexOf('<@' + message.author.id + '>')].id == message.channel.id && Manager.usersGame('<@' + message.author.id + '>').hasRoom(message.toString().substr(0,1) - 1)) {
             if (Manager.usersGame('<@' + message.author.id + '>').turn == 1 + Manager.usersGame('<@' + message.author.id + '>').players.indexOf('<@' + message.author.id + '>')) {
                 Manager.usersGame('<@' + message.author.id + '>').makeMove(message.toString().substr(0,1) - 1, 1 + Manager.usersGame('<@' + message.author.id + '>').players.indexOf('<@' + message.author.id + '>'));
-                message.delete();
+                message.guild && message.delete().catch();
                 console.log(message.author.username + " moved " + message.toString() + ' at ' + new Date().toTimeString().split(' ')[0])
                 return;
             }
